headers/os: Make headers generator-friendly
Make Haiku headers a bit more friendly to binding generators by:
- Giving some `enum`s names (especially those that appear in
default arguments).
- Converting some closely-related constant integers into enums.
- Converting an internal `inline` function into a macro so that
the result could be evaluated in compile time.

+++ src/apps/diskprobe/FindWindow.cpp
@@ -518,7 +518,7 @@ FindWindow::FindWindow(BRect _rect, BMessage& previous, BMessenger& target,
 
 	BRect rect = Bounds().InsetByCopy(5, 5);
 	BMenuField* menuField = new BMenuField(rect, B_EMPTY_STRING,
-		B_TRANSLATE("Mode:"), fMenu, B_FOLLOW_LEFT | B_FOLLOW_TOP);
+		B_TRANSLATE("Mode:"), fMenu, (uint32)(B_FOLLOW_LEFT | B_FOLLOW_TOP));
 	menuField->SetDivider(menuField->StringWidth(menuField->Label()) + 8);
 	menuField->ResizeToPreferred();
 	view->AddChild(menuField);
